Implement Adelaide typechecker for Cheshire
I will begin this commit message by apologizing for how large it is. The
fact of the matter is that the typechecker needs to be written in one
fell swoop, due to the nature of its many interconnecting parts and the
fact that design decisions are typically intertwined in such a way that
iterating in several commits is misleading and confusing.

This commit implements:
- The typechecker
  * Assumption "facts" lowering and simplification
  * Expression lowering and satisfaction
  * Impl selection and trait/method/assoc ty selection
  * Impl goal solving
  * Type unification and inferencing
  (among other small bits and pieces)
- Some changes to lowering to make typechecking easier
- Some refactoring of the cheshire stdlib

Also, fixed a small bug where the Not and Negate operators were flipped